Add transposition option to quantization pass
This PR adds the semantics of transposition to the quantization encoding
function (enabled by default), and it now will transpose the weights
from T to N. By doing this, the linear layers in the model are switched
back to NN matmul.

The PR keeps the no-transposition option for decoding the word embedding
which is queried at the beginning of each timestep.

Fusion passes for "decode + matmul + (optional) ewise" are updated
accordingly.

Given the linear layers are switched back to NN matmul, the matmul
optimization is still on going and will be brought in later.