Allow configuring conversation (#121)
I've added the ability to customize more configuration options, as well
as the conversation options. I couldn't build the project as I had
issues with it looking for `tvm_home`, so I have not tested this and
would like some feedback on if this code makes sense.

I've created this as suggested in
yi @tqchen.