[SLIM] Make web-llm compatible with SLIM models (#219)
Make web-llm runtime compatible with SLIM models, a new workflow in
mlc-llm. This depends on https://github.com/apache/tvm/pull/16198 and is
currently not backward-compatible with non-SLIM model, since we
introduced a new global function to TVM.

The main difference is reading the metadata and loading the parameters.

Besides, we now read stop_tokens from the conversation template instead
of relying on metadata.