[OpenAI] Support seed for reproducible generation (#335)
We support `seed` in `ChatCompletionRequest` so requests' results are
reproducible.

As stated in the docstring in
`src/openai_api_protocols/chat_completion.ts`, seeding is done at a
request level, rather than a choice level. So if a request with `n > 1`
is seeded, the choices would still have different results. But if two
requests, both with `n > 1` and share the same seed, would generate
identical results, across all choices. This is demonstrated in
`examples/openai-api/src/seed.ts`, where we rigorously compare the
strings generated.

Implementation wise, this is achieved with a customized implementation
of linear congruential generator in TVMjs's runtime, since JS's
`Math.random()` does not support seeding.