[ServiceWorker] Reload model when service worker killed (#471)
Service worker may be killed by the browser despite us sending
`keepAlive` messages. Upon killed, all states in the worker thread is
lost. When the frontend sends another request, expecting the model
loaded, the backend MLCEngine throws an error, essentially due to the
mismatch of frontend's expectation and backend's reality.

This PR adds `modelId` and `chatOpts` to `WebWorkerMLCEngine` (hence
ServiceWorker and ExtensionServiceWorker), with the semantics being: the
model and config that the frontend expects the backend to be currently
loaded with. When sending a request, we send the expectation as well.

For backend, upon receiving the chat completion request, it compares the
expectation with its current state. If not matched (due to service
worker killed), we reload before carrying out the request.

---------

Co-authored-by: Nestor Qin <imba.qxy@gmail.com>