[Version][Trivial] Bump version to 0.2.47 (#498)
Exactly the same as 0.2.46 except we update TVMjs to
https://github.com/apache/tvm/commit/32e9a48b1fbc78a58447f392f2530c98553eb3dc.

The main change is that, when requesting a WebGPU device, we may fall
back `maxBufferSize` to the default 256MB when 1GB is too much for the
device. This is motivated by iOS support on Safari, tested on iOS 18
Developer Beta. Fore more, see:
- https://github.com/apache/tvm/pull/17150

### Model WASM Version
No change, still v0.2.43.

### TVMjs
Updated, compiled at
https://github.com/apache/tvm/commit/32e9a48b1fbc78a58447f392f2530c98553eb3dc