[Worker] Move reload upon worker termination to WebWorker (#533)
In https://github.com/mlc-ai/web-llm/pull/471, we add `modelId` and
`chatOpts` to `ServiceWorkerMLCEngineHandler ` and
`ExtensionServiceWorkerMLCEngineHandler`. The motivation was to reload
the model once the handler detects an out-of-sync between what the
frontend engine expects versus what the low-level engine is loaded with.
This out-of-sync usually happens when a service worker is killed
unexpectedly.

Recently, it is observed that such out-of-sync can happen in web worker
as well. Therefore, we move the logic to `WebWorkerMLCEngineHandler`.
The aforementioned handlers calls `super.onmessage()` at the end, so no
change in behavior is expected from them.

Tested in WebLLMChat, intentionally terminating service worker and
sending another message trigger the same behavior as before.