[OpenAI] Support text completion via engine.completions.create() (#534)
`completions.create()`, as opposed to `chat.completions.create()`, is
something we have not supported prior to this PR. Compared to
`chat.completions`, `completions` is pure text completion with no
conversation. That is, given the user's input prompt, the model
autoregressively generates, ignoring any chat template. For more, see
`examples/text-completion` and
https://platform.openai.com/docs/api-reference/completions/object