[Model] Deliver Qwen2.5-Coder 0.5B and 3B (#627)
This PR delivers Qwen2.5-Coder 0.5B and 3B, as 1.5B and 7B are already
supported. We add the following models to the model list:
- `Qwen2.5-Coder-3B-Instruct-q4f16_1-MLC`
- `Qwen2.5-Coder-3B-Instruct-q4f32_1-MLC`
- `Qwen2.5-Coder-0.5B-Instruct-q4f16_1-MLC`
- `Qwen2.5-Coder-0.5B-Instruct-q4f32_1-MLC`
- `Qwen2.5-Coder-0.5B-Instruct-q0f16-MLC`
- `Qwen2.5-Coder-0.5B-Instruct-q0f32-MLC`

Note that the official `config.json` of the coder models are identical
to that of the chat models, so we do not need to compile new WASMs.